requirements.txt
tensorflow
numpy
Pillow
h5py
spicy